\name{aircraftR}
\docType{data}
\alias{aircraftR}
\title{Correlations between characteristics of aircraft}
\description{
   Correlations between SPR (specific power), RGF (flight range factor),
   PLF (payload) and SLF (sustained load factor) for 21 types of aircraft.
}
\usage{data(aircraftR)}
\format{a matrix containing the correlations}
\source{Gower and Hand, Table 2.1}
\references{Gower, J.C. and Hand, D.J. (1996) \emph{Biplots}, Chapman & Hall, London}
\keyword{datasets}
     