\name{artificialR}
\docType{data}
\alias{artificialR}
\title{Correlations for 10 generated variables}
\description{
   A 10 by 10 artificial correlation matrix
}
\usage{data(artificialR)}
\format{A matrix of correlations}
\source{Trosset (2005), Table 1.}
\references{
   Trosset, M.W. (2005) Visualizing correlation. \emph{Journal of Computational and Graphical Statistics},
   14(1), pp. 1--19.
}
\keyword{datasets}
     