\name{athletesR}
\docType{data}
\alias{athletesR}
\title{Correlation matrix of characteristics of Australian athletes}
\description{
   Correlation matrix of 12 characteristics of Austration athletes (Sex,
   Height, Weight, Lean Body Mass, RCC, WCC, Hc, Hg, Ferr, BMI, SSF, Bfat)
}
\usage{data(athletesR)}
\format{A matrix of correlations}
\source{Weisberg (2005), file ais.txt}
\references{
   Weisberg, S. (2005) \emph{Applied Linear Regression}. Third edition, John Wiley \& Sons,
   New Jersey.
}
\keyword{datasets}
     