\name{berkeleyR}
\docType{data}
\alias{berkeleyR}
\title{Correlation matrix for boys of the Berkeley Guidance Study}
\description{
   Correlation matrix for sex, height and weight at age 2, 9 and 18 and somatotype
}
\usage{data(berkeleyR)}
\format{A matrix of correlations}
\source{Weisberg (2005), file BGSBoys.txt}
\references{
   Weisberg, S. (2005) \emph{Applied Linear Regression}. Third edition, John Wiley & Sons,
   New Jersey.
}
\keyword{datasets}
     