\name{cathedralsR}
\docType{data}
\alias{cathedralsR}
\title{Correlation matrix for height and length}
\description{Correlation between nave height and total length}
\usage{data(cathedralsR)}
\format{A matrix of correlations}
\source{Weisberg (2005), file cathedral.txt}
\references{
   Weisberg, S. (2005) \emph{Applied Linear Regression}. Third edition, John Wiley \& Sons,
   New Jersey.
}
\keyword{datasets}
