\name{countriesR}
\docType{data}
\alias{countriesR}
\title{Correlations between educational and demographic variables}
\description{Correlations between infant mortality, educational and demographic variables
  (infd, phys, dens, agds, lit, hied, gnp)
}
\usage{data(countriesR)}
\format{A matrix of correlations}
\source{Chatterjee and Hadi (1988)}
\references{
   Chatterjee, S. and Hadi, A.S. (1988), \emph{Sensitivity Analysis in Regression}.
   Wiley, New York.
}
\keyword{datasets}
     