\name{fysiologyR}
\docType{data}
\alias{fysiologyR}
\title{Correlations between thirtheen fysiological variables}
\description{
   Correlations of 13 fysiological variables (sys, dia, p.p., pul, cort,
   u.v., tot/100, adr/100, nor/100, adr/tot, tot/hr, adr/hr, nor/hr)
   obtained from 48 medical students
}
\usage{data(fysiologyR)}
\format{A matrix of correlations}
\source{Hills (1969), Table 1.}
\references{
   Hills, M (1969) On looking at large correlation matices
   \emph{Biometrika} 56(2): pp. 249.
}
\keyword{datasets}
     