\name{gobletsR}
\docType{data}
\alias{gobletsR}
\title{Correlations between size measurements of archeological goblets}
\description{Correlations between 6 size measurements of archeological goblets}
\usage{data(gobletsR)}
\format{A matrix of correlations}
\source{Manly (1989)}
\references{
   Manly, B.F.J. (1989) \emph{Multivariate statistical methods: a
   primer}. Chapman and Hall, London.
}
\keyword{datasets}
     