\name{PearsonLee}
\docType{data}
\alias{PearsonLee}
\title{Heights of mothers and daughters}
\description{
   Heights of 1375 mothers and daughters (in cm) in the UK in 1893-1898. 
}
\usage{data(PearsonLee)}
\format{dataframe with Mheight and Dheight}
\source{Weisberg, Chapter 1}
\references{Weisberg, S. (2005) \emph{Applied Linear Regression}, John
  Wiley & Sons, New Jersey}
\keyword{datasets}
     