\name{proteinsR}
\docType{data}
\alias{proteinsR}
\title{Correlations between sources of protein} 
\description{
  Correlations between sources of protein for a number of countries (Red
  meat, White meat, Eggs, Milk, Fish, Cereals, Starchy food, Nuts,
  Fruits and vegetables.
}
\usage{data(proteinR)}
\format{A matrix of correlations}
\source{Manly (1989)}
\references{
   Manly, B.F.J. (1989) \emph{Multivariate statistical methods: a
   primer}. Chapman and Hall, London.
}
\keyword{datasets}
     