\name{recordsR}
\docType{data}
\alias{recordsR}
\title{Correlations between  national track records for men}
\description{
   Correlations between national track records for men
   (100m,200m,400m,800m,1500m,5000m,10.000m and Marathon
}
\usage{data(recordsR)}
\format{A matrix of correlations}
\source{Johnson and Wichern, Table 8.6}
\references{
   Johnson, R.A. and Wichern, D.W. (2002) \emph{Applied Multivariate
     Statistical Analysis.} Fifth edition. New Jersey: Prentice Hall.
}
\keyword{datasets}
     