\name{storksR}
\docType{data}
\alias{storksR}
\title{Correlations between three  variables}
\description{
   Danish data from 1953-1977 giving the correlations between nesting storks,
   human birth rate and per capita electricity consumption.
}
\usage{data(storksR)}
\format{A matrix of correlations}
\source{Gabriel and Odoroff, Table 1.}
\references{
   Gabriel, K. R. and Odoroff, C. L. (1990) Biplots in biomedical research.
   \emph{Statistics in Medicine} 9(5): pp. 469-485.
}
\keyword{datasets}
     