\name{students}
\docType{data}
\alias{students}
\title{Marks for 5 student exams}
\description{
   Matrix of marks for five exams, two with closed books and
   three with open books (Mechanics (C), Vectors (C),
   Algebra (O), Analysis (O) and Statistics (O)).    
}
\usage{data(students)}
\format{A data matrix}
\source{Mardia et al., Table 1.2.1}
\references{
   Mardia, K.V., Kent, J.T. and Bibby, J.M. (1979) \emph{Multivariate
   Analysis}, Academic Press London.
}
\keyword{datasets}
     