\name{studentsR}
\docType{data}
\alias{studentsR}
\title{Correlations between marks for 5 exams}
\description{
   Correlation matrix of marks for five exams, two with closed books and
   three with open books (Mechanics (C), Vectors (C),
   Algebra (O), Analysis (O) and Statistics (O)).    
}
\usage{data(studentsR)}
\format{A matrix of correlations}
\source{Mardia et al., Table 1.2.1}
\references{
   Mardia, K.V., Kent, J.T. and Bibby, J.M. (1979) \emph{Multivariate
   Analysis}, Academic Press London.
}
\keyword{datasets}
     